Registry: hold registry cache for a single thread only (HSC-1035)
We destroy the connection cache when we discover that we're in
a new thread, to avoid using a SQLite connection from a thread
other than which it was created in (resulting in: "ProgrammingError:
SQLite objects created in a thread can only be used in that same
thread.").  multiprocessing runs a thread within each process
(only one at a time), and different iterations in a process appear
to use different threads.

This is not attempting to be thread-safe (this is inherently NOT
thread-safe --- there is no locking or thread-specific data), but
to support multiprocessing, as SQLite (and possibly others) will not
use a connection created in a different thread.

Could have chosen to hold the cache of connections indexed on the
thread identifier, but that means several connections for the same
database could be open at the same time, which defeats the purpose
of the cache (or, at least, makes it less effective).